pyproject.toml
[tool.poetry]
name = "toll-plaza-simulator"
version = "0.1.0"
description = ""
authors = ["Kimba SABI N'GOYE <kimbasabingoye@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
pika = "^1.3.2"
google-cloud-pubsub = "^2.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
